Merge homolog para master

+++ src/main/java/org/financeiro/dtos/CategoriaDto.java
@@ -1,6 +1,6 @@
 package org.financeiro.dtos;
 
-public record CategoriaDto(int id,
+public record CategoriaDto(long id,
                            String nome,
                            String descricao,
                            UsuarioDto usuarioDto) {
